P2/P3 code quality follow-up: manifest split, bounds docs, readiness
- Rename scripts/tmp-discover-stat-mod-plugs.ts → discover-stat-mod-plugs.ts
- Add scripts/README.md with script index
- Extract plug categorization loop to derive-plug-lookups.ts (derive.ts 658→472)
- Document bounds API split: maxAchievable*Stat for gray bands, maxFeasibleStatTarget for slider-min search
- Update bounds-heuristic JSDoc and gray-band handoff doc
- Single auto-run readiness source: hook exports liveAutoRunReadiness for placeholder UI
